Homework JUnit/TestNG - part3
