+++ .idea/workspace.xml
@@ -4,13 +4,15 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="7c307568-bef4-4215-ac92-f8bc985d133c" name="Changes" comment="Homework JUnit/TestNG">
+    <list default="true" id="7c307568-bef4-4215-ac92-f8bc985d133c" name="Changes" comment="Homework JUnit/TestNG - part3">
+      <change afterPath="$PROJECT_DIR$/src/test/java/familytest/WomanTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rd/family/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rd/family/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rd/family/Man.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rd/family/Man.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/com/rd/family/Person.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rd/family/Person.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/com/rd/family/Woman.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rd/family/Woman.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/familytest/PersonsTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/familytest/PersonsTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/familytest/RelationshipTests.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/familytest/RelationshipTests.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/familytest/PersonsTest.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/java/familytest/RelationshipTests.java" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/test/resources/regression.xml" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/resources/regression.xml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -58,7 +60,7 @@
       <recent name="D:\Studying\QA Automation\Homework\dz-11\src\test" />
     </key>
   </component>
-  <component name="RunManager" selected="TestNG.RelationshipTests.testRegisterPartnership">
+  <component name="RunManager" selected="TestNG.WomanTest">
     <configuration name="D:/Studying/QA Automation/Homework/dz-11/src/test/resources/regression.xml (1)" type="TestNG" temporary="true" nameIsGenerated="true">
       <module name="dz-11" />
       <option name="SUITE_NAME" value="$PROJECT_DIR$/src/test/resources/regression.xml" />
@@ -69,24 +71,7 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="PersonsTest" type="TestNG" temporary="true" nameIsGenerated="true">
-      <module name="APITest" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="familytest.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <option name="PACKAGE_NAME" value="familytest" />
-      <option name="MAIN_CLASS_NAME" value="familytest.PersonsTest" />
-      <option name="TEST_OBJECT" value="CLASS" />
-      <properties />
-      <listeners />
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
-    <configuration name="RelationshipTests (1)" type="TestNG" temporary="true" nameIsGenerated="true">
+    <configuration name="ManTest" type="TestNG" temporary="true" nameIsGenerated="true">
       <module name="dz-11" />
       <extension name="coverage">
         <pattern>
@@ -95,7 +80,7 @@
         </pattern>
       </extension>
       <option name="PACKAGE_NAME" value="familytest" />
-      <option name="MAIN_CLASS_NAME" value="familytest.RelationshipTests" />
+      <option name="MAIN_CLASS_NAME" value="familytest.ManTest" />
       <option name="TEST_OBJECT" value="CLASS" />
       <properties />
       <listeners />
@@ -103,8 +88,8 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="RelationshipTests" type="TestNG" temporary="true" nameIsGenerated="true">
-      <module name="APITest" />
+    <configuration name="PersonTest" type="TestNG" temporary="true" nameIsGenerated="true">
+      <module name="dz-11" />
       <extension name="coverage">
         <pattern>
           <option name="PATTERN" value="familytest.*" />
@@ -112,7 +97,7 @@
         </pattern>
       </extension>
       <option name="PACKAGE_NAME" value="familytest" />
-      <option name="MAIN_CLASS_NAME" value="familytest.RelationshipTests" />
+      <option name="MAIN_CLASS_NAME" value="familytest.PersonTest" />
       <option name="TEST_OBJECT" value="CLASS" />
       <properties />
       <listeners />
@@ -138,13 +123,30 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
+    <configuration name="WomanTest" type="TestNG" temporary="true" nameIsGenerated="true">
+      <module name="dz-11" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="familytest.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="familytest" />
+      <option name="MAIN_CLASS_NAME" value="familytest.WomanTest" />
+      <option name="TEST_OBJECT" value="CLASS" />
+      <properties />
+      <listeners />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
     <recent_temporary>
       <list>
-        <item itemvalue="TestNG.RelationshipTests.testRegisterPartnership" />
+        <item itemvalue="TestNG.WomanTest" />
+        <item itemvalue="TestNG.PersonTest" />
+        <item itemvalue="TestNG.ManTest" />
         <item itemvalue="TestNG.D:/Studying/QA Automation/Homework/dz-11/src/test/resources/regression.xml (1)" />
-        <item itemvalue="TestNG.RelationshipTests (1)" />
-        <item itemvalue="TestNG.RelationshipTests" />
-        <item itemvalue="TestNG.PersonsTest" />
+        <item itemvalue="TestNG.RelationshipTests.testRegisterPartnership" />
       </list>
     </recent_temporary>
   </component>
