+++ src/main/java/com/rd/family/Man.java
@@ -1,9 +1,8 @@
 package com.rd.family;
 public class Man extends Person{
-    public Man (String firstName,String lastName,Integer age,String partner,String profession){
-        super(firstName,lastName,age,partner,profession);
+    public Man (String firstName,String lastName,Integer age,String partner,String profession, String partnerLastName, String maidenName){
+        super(firstName,lastName,age,partner,profession, partnerLastName,maidenName );
     }
-
 
     @Override
     public String getPartnerLastName() {
