+++ src/test/resources/regression.xml
@@ -1,15 +1,15 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Family regression test">
-
+<groups>
+    <run>
+        <include name="AllTests"/>
+        <include name="GettersSetters"/>
+    </run>
+</groups>
     <test name="Family regression">
-        <classes>
-            <class name="familytest.PersonsTest"/> <!-- only getters and setters-->
-            <class name="familytest.RelationshipTests"/> <!-- all tests-->
-        </classes>
-<!--        <packages>-->
-<!--            <package name="familytest"/>-->
-<!--        </packages>-->
-
+        <packages>
+            <package name="familytest"/>
+        </packages>
 
     </test>
